Link against OneCore.lib to be Win10 compliant
From MSDN:
For convenience, an umbrella library named OneCore.lib is provided in the
Microsoft Windows Software Development Kit (SDK), which provides the
exports for the subset of Win32 APIs that are common to all Windows 10
devices.

If you link your app or driver to OneCore.lib, and you only call Win32
APIs in that library, then your app or driver will load successfully on
all Windows 10 devices.

https://msdn.microsoft.com/en-us/library/windows/desktop/mt657574(v=vs.85).aspx